--- user/src/components/VenueApplication/hooks/index.ts
@@ -1,7 +1,0 @@
-import { Place } from '@/api/venueApplication';
-
-export const convertPlacesToOptions = (places: Place[]) => {
-  return places.map((place) => {
-    return { id: place.id, name: place.name };
-  });
-};
